*: move opening sst writer to importer worker threads
