+++ tests/failpoints/cases/mod.rs
@@ -5,6 +5,7 @@ mod test_conf_change;
 mod test_coprocessor;
 mod test_early_apply;
 mod test_gc_worker;
+mod test_import_service;
 mod test_merge;
 mod test_pd_client;
 mod test_pending_peers;
