+++ tests/integrations/import/test_sst_service.rs
@@ -6,51 +6,24 @@ use std::time::Duration;
 
 use futures::{stream, Future, Stream};
 use tempfile::Builder;
-use uuid::Uuid;
 
-use grpcio::{ChannelBuilder, Environment, Result, WriteFlags};
 use kvproto::import_sstpb::*;
-use kvproto::kvrpcpb::*;
-use kvproto::tikvpb::*;
 
 use pd_client::PdClient;
-use test_raftstore::*;
 use test_sst_importer::*;
-use tikv_util::HandyRwLock;
 
-const CLEANUP_SST_MILLIS: u64 = 10;
+use super::util::*;
 
-fn new_cluster() -> (Cluster<ServerCluster>, Context) {
-    let count = 1;
-    let mut cluster = new_server_cluster(0, count);
-    let cleanup_interval = Duration::from_millis(CLEANUP_SST_MILLIS);
-    cluster.cfg.raft_store.cleanup_import_sst_interval.0 = cleanup_interval;
-    cluster.run();
-
-    let region_id = 1;
-    let leader = cluster.leader_of_region(region_id).unwrap();
-    let epoch = cluster.get_region_epoch(region_id);
-    let mut ctx = Context::default();
-    ctx.set_region_id(region_id);
-    ctx.set_peer(leader);
-    ctx.set_region_epoch(epoch);
-
-    (cluster, ctx)
-}
-
-fn new_cluster_and_tikv_import_client(
-) -> (Cluster<ServerCluster>, Context, TikvClient, ImportSstClient) {
-    let (cluster, ctx) = new_cluster();
-
-    let ch = {
-        let env = Arc::new(Environment::new(1));
-        let node = ctx.get_peer().get_store_id();
-        ChannelBuilder::new(env).connect(cluster.sim.rl().get_addr(node))
-    };
-    let tikv = TikvClient::new(ch.clone());
-    let import = ImportSstClient::new(ch);
-
-    (cluster, ctx, tikv, import)
+macro_rules! assert_to_string_contains {
+    ($e:expr, $substr:expr) => {{
+        let msg = $e.to_string();
+        assert!(
+            msg.contains($substr),
+            "msg: {}; expr: {}",
+            msg,
+            stringify!($e)
+        );
+    }};
 }
 
 #[test]
@@ -271,99 +244,4 @@ fn test_ingest_sst_region_not_found() {
     ingest.set_sst(meta);
     let resp = import.ingest(&ingest).unwrap();
     assert!(resp.get_error().has_region_not_found());
-}
-
-fn new_sst_meta(crc32: u32, length: u64) -> SstMeta {
-    let mut m = SstMeta::default();
-    m.set_uuid(Uuid::new_v4().as_bytes().to_vec());
-    m.set_crc32(crc32);
-    m.set_length(length);
-    m
-}
-
-fn send_upload_sst(
-    client: &ImportSstClient,
-    meta: &SstMeta,
-    data: &[u8],
-) -> Result<UploadResponse> {
-    let mut r1 = UploadRequest::default();
-    r1.set_meta(meta.clone());
-    let mut r2 = UploadRequest::default();
-    r2.set_data(data.to_vec());
-    let reqs: Vec<_> = vec![r1, r2]
-        .into_iter()
-        .map(|r| (r, WriteFlags::default()))
-        .collect();
-    let (tx, rx) = client.upload().unwrap();
-    let stream = stream::iter_ok(reqs);
-    stream.forward(tx).and_then(|_| rx).wait()
-}
-
-fn send_write_sst(
-    client: &ImportSstClient,
-    meta: &SstMeta,
-    keys: Vec<Vec<u8>>,
-    values: Vec<Vec<u8>>,
-    commit_ts: u64,
-) -> Result<WriteResponse> {
-    let mut r1 = WriteRequest::default();
-    r1.set_meta(meta.clone());
-    let mut r2 = WriteRequest::default();
-
-    let mut batch = WriteBatch::default();
-    let mut pairs = vec![];
-
-    for (i, key) in keys.iter().enumerate() {
-        let mut pair = Pair::default();
-        pair.set_key(key.to_vec());
-        pair.set_value(values[i].to_vec());
-        pairs.push(pair);
-    }
-    batch.set_commit_ts(commit_ts);
-    batch.set_pairs(pairs.into());
-    r2.set_batch(batch);
-
-    let reqs: Vec<_> = vec![r1, r2]
-        .into_iter()
-        .map(|r| (r, WriteFlags::default()))
-        .collect();
-
-    let (tx, rx) = client.write().unwrap();
-    let stream = stream::iter_ok(reqs);
-    stream.forward(tx).and_then(|_| rx).wait()
-}
-
-fn check_ingested_kvs(tikv: &TikvClient, ctx: &Context, sst_range: (u8, u8)) {
-    for i in sst_range.0..sst_range.1 {
-        let mut m = RawGetRequest::default();
-        m.set_context(ctx.clone());
-        m.set_key(vec![i]);
-        let resp = tikv.raw_get(&m).unwrap();
-        assert!(resp.get_error().is_empty());
-        assert!(!resp.has_region_error());
-        assert_eq!(resp.get_value(), &[i]);
-    }
-}
-
-fn check_ingested_txn_kvs(tikv: &TikvClient, ctx: &Context, sst_range: (u8, u8), start_ts: u64) {
-    for i in sst_range.0..sst_range.1 {
-        let mut m = GetRequest::default();
-        m.set_context(ctx.clone());
-        m.set_key(vec![i]);
-        m.set_version(start_ts);
-        let resp = tikv.kv_get(&m).unwrap();
-        assert!(!resp.has_region_error());
-        assert_eq!(resp.get_value(), &[i]);
-    }
-}
-
-fn check_sst_deleted(client: &ImportSstClient, meta: &SstMeta, data: &[u8]) {
-    for _ in 0..10 {
-        if send_upload_sst(client, meta, data).is_ok() {
-            // If we can upload the file, it means the previous file has been deleted.
-            return;
-        }
-        thread::sleep(Duration::from_millis(CLEANUP_SST_MILLIS));
-    }
-    send_upload_sst(client, meta, data).unwrap();
 }
